pyproject.toml
[project]
name = "myfilehelper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
	"Flask==3.0.3",
	"Flask-SocketIO==5.3.6",
	"python-socketio==5.11.3",
	"PyYAML==6.0.2",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.2,<9",
	"pytest-cov>=5,<6",
	"ruff>=0.6,<1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --import-mode=importlib"
pythonpath = ["."]

[tool.ruff]
target-version = "py39"
line-length = 120
